Played around with a new heuristic search
it's not really a new greedy algorithm at all. It has forwards and backw
ards moves. It will improve things in some cases, but it needs some pret
ty thorough testing to see if it's actually useful though. It also adds
duplicate schemes at the moment, which isn't a problem but is a bit sill
y.