example of how clustering might work
This is more a note to self than useful code, but it's a start. Have a
look at the notes. The basic idea is that this could be a more
efficient but less guaranteed way to estimate schemes, especially for
large datasets. For really large datasets we should also just not
bother to estimate the scheme likelihoods in PF, PhyML gets too slow on
really big datasets. There are two options here:

1. EASY: output a bunch of raxml partition.txt files for people to
implement themsleves. This is probably best since for really big
datasets we're unlikely to be able to implement the best version of
RaxML (i.e. some big server type version).

2. HARDER: package it all up with RaxML so that we can do that second
part ourselves. It might be best in this instance to fix the topology
to the NJ topology as before. RaxML should be pretty quick then... The
main pain here would be getting the right collection of RaxML
executables.