added raxml mac executable
This is just the basic RAxML v 7.3.1, and only the mac one.
Also note that I've just named it 'raxml', because there's no
particular reason why people shouldn't use any version of raxml. The
faster and more specific the better. The basic version is easiest to
distribute.

Finally, windows version should have .exe when we add it